requirements.txt
akshare==1.15.87
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
DateTime==5.5
daytime==0.4
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
efinance==0.5.2
et_xmlfile==2.0.0
executing==2.2.0
filterpy==1.4.5
fonttools==4.56.0
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
jiter==0.8.2
jsonpath==0.82.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lxml==5.3.1
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mini-racer==0.12.4
mplfinance==0.12.10b0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.63.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
py==1.11.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pyzmq==26.2.1
requests==2.32.3
retry==0.9.2
rich==13.9.4
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tabulate==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
volcengine-python-sdk==1.0.124
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
xlrd==2.0.1
yfinance==0.2.54
zope.interface==7.2
